+++ deployment/library/purge_oldest_files.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/python3
 import os
 import os.path
 import traceback